+++ packages/core/src/manglers/html-attributes.ts
@@ -1,3 +1,6 @@
+import type { CharSet } from "../types";
+
+import { ALL_LOWERCASE_CHARS, ALL_NUMBER_CHARS } from "../characters";
 import SimpleManglerPlugin from "./utils/simple-mangler.class";
 
 /**
@@ -153,6 +156,26 @@ export default class HtmlAttributeMangler extends SimpleManglerPlugin {
   static readonly _ID = "html-attribute-mangler";
 
   /**
+   * The list of reserved strings that are always reserved because they are
+   * illegal HTML attribute names.
+   *
+   * @since v0.1.7
+   */
+  static readonly ALWAYS_RESERVED: string[] = ["([0-9]|-|_).*"];
+
+  /**
+   * The character set used by {@link HtmlAttributeMangler}. Note that HTML
+   * attributes are case insensitive, so only lowercase letters are used.
+   *
+   * @since v0.1.7
+   */
+  static readonly CHARACTER_SET: CharSet = [
+    ...ALL_LOWERCASE_CHARS,
+    ...ALL_NUMBER_CHARS,
+    "-", "_",
+  ];
+
+  /**
    * The default patterns used by a {@link HtmlAttributeMangler}.
    *
    * @since v0.1.0
@@ -181,6 +204,7 @@ export default class HtmlAttributeMangler extends SimpleManglerPlugin {
    */
   constructor(options: HtmlAttributeManglerOptions={}) {
     super(HtmlAttributeMangler._ID, {
+      charSet: HtmlAttributeMangler.CHARACTER_SET,
       patterns: HtmlAttributeMangler.getPatterns(options.attrNamePattern),
       reserved: HtmlAttributeMangler.getReserved(options.reservedAttrNames),
       prefix: HtmlAttributeMangler.getPrefix(options.keepAttrPrefix),
@@ -210,11 +234,12 @@ export default class HtmlAttributeMangler extends SimpleManglerPlugin {
    * @returns The reserved names to be used.
    */
   private static getReserved(reservedAttrNames?: string[]): string[] {
-    if (reservedAttrNames === undefined) {
-      return HtmlAttributeMangler.DEFAULT_RESERVED;
+    let configured = reservedAttrNames;
+    if (configured === undefined) {
+      configured = HtmlAttributeMangler.DEFAULT_RESERVED;
     }
 
-    return reservedAttrNames;
+    return HtmlAttributeMangler.ALWAYS_RESERVED.concat(configured);
   }
 
   /**