+++ packages/core/src/manglers/utils/simple-mangler.class.ts
@@ -1,4 +1,4 @@
-import type { MangleEngine, ManglerFile } from "../../types";
+import type { CharSet, MangleEngine, ManglerFile } from "../../types";
 
 import BaseManglerPlugin from "./base-mangler.class";
 
@@ -8,6 +8,13 @@ import BaseManglerPlugin from "./base-mangler.class";
  * @since v0.1.0
  */
 interface SimpleManglerOptions {
+  /**
+   * The character set to use when mangling.
+   *
+   * @since v0.1.7
+   */
+  charSet: CharSet;
+
   /**
    * One or more patterns that should be mangled.
    *
@@ -34,7 +41,7 @@ interface SimpleManglerOptions {
  * The {@link SimpleManglerPlugin} abstract class provides an implementation of
  * a {@link WebManglerPlugin} that deals with the handling of {@link
  * WebManglerLanguagePlugin} and implements {@link WebManglerPlugin.mangle} for
- * a given set of patterns, reserved values, and prefix.
+ * a given character set, set of patterns, reserved values, and prefix.
  *
  * It is recommended to extend this class - or {@link BaseManglerPlugin} or
  * {@link MultiMangler}, depending on your needs - if you're implementing a
@@ -43,6 +50,11 @@ interface SimpleManglerOptions {
  * @since v0.1.0
  */
 export default abstract class SimpleManglerPlugin extends BaseManglerPlugin {
+  /**
+   * The character set to use when mangling.
+   */
+  private readonly charSet: CharSet;
+
   /**
    * The pattern(s) to be mangled.
    */
@@ -62,18 +74,19 @@ export default abstract class SimpleManglerPlugin extends BaseManglerPlugin {
    * Initialize a new {@link WebManglerPlugin}.
    *
    * @param id The identifier for the {@link WebManglerPlugin}.
-   * @param options The pattern, reserved values, and prefix.
+   * @param options The character set, patterns, reserved values, and prefix.
    */
   constructor(id: string, options: SimpleManglerOptions) {
     super(id);
+    this.charSet = options.charSet;
     this.patterns = options.patterns;
     this.reserved = options.reserved;
     this.prefix = options.prefix;
   }
 
   /**
-   * Mangle the `files` with the configured pattern, reserved values, and
-   * prefix.
+   * Mangle the `files` with the configured character set, pattern, reserved
+   * values, and prefix.
    *
    * @inheritDoc
    * @since v0.1.0
@@ -87,6 +100,7 @@ export default abstract class SimpleManglerPlugin extends BaseManglerPlugin {
       this.expressions,
       this.patterns,
       {
+        charSet: this.charSet,
         reservedNames: this.reserved,
         manglePrefix: this.prefix,
       },